Genome view
* I created a `genomes view` to test how the “genome-card” would look
like. Although there is no links available to access this testing view,
I have make it work in order to be able to start testing as well
* The unit test was added earlier
* Since this is a very temporal strategy, I am including the css in
`main.css` by now

Not sure yet though how this will work. But I want to keep it there. It
could be a directive and pass a bunch of parameters, or it could be
something different. Times will decide…