Unit testing: toward testing templateUrl
When the template of directives is given through templateUrl is very
tricky. Here I am making some tests to see how it has to be tested. I
don’t get it yet.
One of the approaches suggested online is to use
karma-ng-html2js-preprocessor to avoid problems with mocking the http
service, and them to add to the karma file the location of the
templates. I don’t know, it does not seem to work so far.