Scripting the `base` tag
The `$locationProvider` module configures how the application deep
linking paths are stored. Requires the tag `<base href=“”> tag`.
Alternatively it could be deactivated  by setting the `requireBase:
false` to the `$locationProvider`. The problem is that removing the
requirement for a tag will have adverse side effects when resolving
relative paths with `$location` in IE9, so it’s better to keep the
default option (`requireBase: true`)