Configuration of dynamic routes
The idea is to use all the old html as templates for the `ng-view`, but
without the need of specifying the route for each of them at the
`$routeProvider`. To do that, a “group name” is pass with the `colon :`

A controller (`MediaController’) is also included just in case is
needed later on.