Adding filter by future events
This filter already implemented was not fully functional and not
working for certain browsers. It requires to convert both dates to the
exact same javascript date format, which it is done here by injecting
filter in the controller. It solves the problem in all the browser
(more sensitive to the data formatting)

#120